+++ webhook/credentials_test.go
@@ -3,6 +3,7 @@ package webhook_test
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 	"net/http"
 
 	. "github.com/onsi/ginkgo/v2"
@@ -10,11 +11,15 @@ import (
 	"gomodules.xyz/jsonpatch/v2"
 	admissionv1 "k8s.io/api/admission/v1"
 	corev1 "k8s.io/api/core/v1"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
+	infra "sigs.k8s.io/cluster-api-provider-gcp/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
+	"github.com/giantswarm/workload-identity-operator-gcp/controllers"
+	gke "github.com/giantswarm/workload-identity-operator-gcp/pkg/gke/membership"
 	"github.com/giantswarm/workload-identity-operator-gcp/webhook"
 )
 
@@ -25,8 +30,12 @@ var _ = Describe("Credentials", func() {
 
 		pod            corev1.Pod
 		serviceAccount *corev1.ServiceAccount
+		gcpCluster     *infra.GCPCluster
 		request        admission.Request
 		response       admission.Response
+
+		gcpProject           = "testing-1234"
+		workloadIdentityPool = fmt.Sprintf("%s.svc.id.goog", gcpProject)
 	)
 
 	BeforeEach(func() {
@@ -41,8 +50,7 @@ var _ = Describe("Credentials", func() {
 				Name:      "the-service-account",
 				Namespace: namespace,
 				Annotations: map[string]string{
-					webhook.AnnotationGCPServiceAccount:      "service-account@email",
-					webhook.AnnotationWorkloadIdentityPoolID: "workload-identity-pool-id",
+					controllers.AnnotationGCPServiceAccount: "service-account@email",
 				},
 			},
 		}
@@ -50,7 +58,8 @@ var _ = Describe("Credentials", func() {
 
 		pod = corev1.Pod{
 			ObjectMeta: metav1.ObjectMeta{
-				Name: "the-pod",
+				Name:      "the-pod",
+				Namespace: namespace,
 				Labels: map[string]string{
 					webhook.LabelWorkloadIdentity: "enabled",
 				},
@@ -76,6 +85,17 @@ var _ = Describe("Credentials", func() {
 			},
 		}
 
+		clusterName := "krillin"
+		gcpCluster = &infra.GCPCluster{
+			TypeMeta: metav1.TypeMeta{},
+			ObjectMeta: metav1.ObjectMeta{
+				Name: clusterName,
+			},
+			Spec: infra.GCPClusterSpec{
+				Project: gcpProject,
+			},
+		}
+
 		request = admission.Request{
 			AdmissionRequest: admissionv1.AdmissionRequest{
 				Object:    encodeObject(pod),
@@ -86,6 +106,7 @@ var _ = Describe("Credentials", func() {
 	})
 
 	JustBeforeEach(func() {
+		Expect(ensureMembershipSecretExists(gcpCluster)).To(Succeed())
 		response = credentialsWebhook.Handle(ctx, request)
 	})
 
@@ -141,6 +162,7 @@ var _ = Describe("Credentials", func() {
 
 	It("injects the secret volume", func() {
 		Expect(response.Allowed).To(BeTrue())
+		fmt.Println(response.Patches)
 		Expect(response.Patches).To(ContainElements(
 			jsonpatch.Operation{
 				Operation: "add",
@@ -153,8 +175,8 @@ var _ = Describe("Credentials", func() {
 							"sources": []interface{}{
 								map[string]interface{}{
 									"serviceAccountToken": map[string]interface{}{
-										"path":              webhook.ServiceAccountTokenPath,
-										"audience":          "workload-identity-pool-id",
+										"path":              controllers.ServiceAccountTokenPath,
+										"audience":          workloadIdentityPool,
 										"expirationSeconds": float64(webhook.TokenExpirationSeconds),
 									},
 								},
@@ -164,7 +186,7 @@ var _ = Describe("Credentials", func() {
 										"optional": false,
 										"items": []interface{}{
 											map[string]interface{}{
-												"key":  webhook.SecretKeyGoogleApplicationCredentials,
+												"key":  controllers.SecretKeyGoogleApplicationCredentials,
 												"path": webhook.GoogleApplicationCredentialsJSONPath,
 											},
 										},
@@ -239,20 +261,6 @@ var _ = Describe("Credentials", func() {
 		})
 	})
 
-	When("the service account isn't annotated with the workload identity id", func() {
-		BeforeEach(func() {
-			modified := serviceAccount.DeepCopy()
-			modified.Annotations = map[string]string{}
-			Expect(k8sClient.Patch(ctx, modified, client.MergeFrom(serviceAccount))).To(Succeed())
-		})
-
-		It("denies the request", func() {
-			Expect(response.AdmissionResponse.Allowed).To(BeFalse())
-			Expect(response.Result).NotTo(BeNil())
-			Expect(response.Result.Code).To(Equal(int32(http.StatusForbidden)))
-		})
-	})
-
 	When("the context has been canceled", func() {
 		It("returns a 500 Internal Server Error", func() {
 			canceledCtx, cancel := context.WithCancel(ctx)
@@ -283,4 +291,42 @@ func encodeObject(obj interface{}) runtime.RawExtension {
 	Expect(err).NotTo(HaveOccurred())
 
 	return runtime.RawExtension{Raw: encodedObj}
+}
+
+func ensureMembershipSecretExists(gcpCluster *infra.GCPCluster) error {
+	membershipSecret := &corev1.Secret{}
+	ctx := context.Background()
+
+	err := k8sClient.Get(ctx, client.ObjectKey{
+		Name:      controllers.MembershipSecretName,
+		Namespace: controllers.DefaultMembershipSecretNamespace,
+	}, membershipSecret)
+
+	if k8serrors.IsNotFound(err) {
+		oidcJwks := []byte{}
+
+		membership := gke.GenerateMembership(*gcpCluster, oidcJwks)
+		membershipJson, err := json.Marshal(membership)
+
+		Expect(err).To(BeNil())
+
+		membershipSecret := &corev1.Secret{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      controllers.MembershipSecretName,
+				Namespace: controllers.DefaultMembershipSecretNamespace,
+				Annotations: map[string]string{
+					controllers.AnnoationMembershipSecretCreatedBy: gcpCluster.Name,
+					controllers.AnnotationSecretManagedBy:          controllers.SecretManagedBy,
+				},
+			},
+			StringData: map[string]string{
+				controllers.SecretKeyGoogleApplicationCredentials: string(membershipJson),
+			},
+		}
+		err = k8sClient.Create(context.Background(), membershipSecret)
+
+		return err
+	}
+
+	return err
 }