+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- GCP cluster reconciler that creates a `secret` with the details of the workload identity membership on a workload cluster
+
 ## [0.3.2] - 2022-08-23
 
 ### Fixed